pyproject.toml
[tool.poetry]
name = "vending-machine-api"
version = "0.1.0"
description = ""
authors = ["mohamed ibrahim"]
readme = "README.md"
packages = [{include = "vending_machine_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-rest-knox = "^4.2.0"
djangorestframework = "^3.14.0"
django-model-utils = "^4.3.1"
drf-access-policy = "^1.3.0"
pre-commit = "^2.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
